Issue #99: Refactored `Match` as es6 module.
