+++ bw_matchbox/assets/modules/Match/@types/TMatchData.d.ts
@@ -1,0 +1,7 @@
+interface TMatchData {
+  id: number; // 4604
+  location: string; // 'United States'
+  name: string; // 'Home furniture - Cabinets and non-wood, nonupholstered; at manufacturer'
+  referenceProduct?: string; // ???
+  unit: string; // ''
+}
