+++ bw_matchbox/assets/modules/Match/@types/TMatchProcess.d.ts
@@ -1,0 +1,6 @@
+interface TMatchProcess {
+  comment: string; // ''
+  location: string; // 'United States'
+  name: string; // 'Home furniture - Cabinets and non-wood, nonupholstered; at manufacturer'
+  unit: string; // 'USD'
+}
