+++ bw_matchbox/assets/modules/Match/@types/TSharedParams.d.ts
@@ -1,0 +1,9 @@
+interface TSharedParams {
+  database: string; // 'USEEIO-copy'
+  initialQueryString: string; // 'home furniture - cabinets and non-wood, nonupholstered; at manufacturer United States'
+  initialTableRows: TMatchData[]; // [...]
+  originalProcess: TMatchProcess; // {name: 'Home furniture - Cabinets and non-wood, nonupholstered; at manufacturer', location: 'United States', unit: 'USD', comment: ''}
+  searchUrlPrefix: string; // '/search/'
+  /** Process name */
+  source: string; // '232'
+}
