+++ bw_matchbox/assets/modules/Match/@types/TTableCell.d.ts
@@ -1,0 +1,5 @@
+interface TTableCell {
+  id: string;
+  value?: string;
+  text?: string;
+}
