+++ bw_matchbox/assets/templates/compare.html
@@ -103,7 +103,7 @@
      */
   };
 
-  console.log('[compare:start]', sharedParams);
+  {# console.log('[compare:start]', sharedParams); #}
 
   // Export to global scope (to access from generated html code handlers).
   window.CompareCore = CompareCore;
