+++ bw_matchbox/assets/templates/index.html
@@ -131,7 +131,7 @@
   };
 
   // DEBUG
-  console.log('[index]', sharedParams);
+  {# console.log('[index]', sharedParams); #}
 
   // Export to global scope (to access from generated html code handlers).
   window.ProcessesList = ProcessesList;
